RFFC5071: Rename rffc5071_drv->rffc5071_spi.
